// frontend/src/sections/prescription-orders/list/createPrescriptionOrder.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../api/graphql";

export type InsertPrescriptionOrderVariables = {
  storeId: string;
  input: {
    id: string;
    patientId: string;
    clinicianId?: string | null;
    diagnosisId?: string | null;
    programId?: string | null;
    prescriptionDatetime?: string | null;
  };
};

export type InsertPrescriptionOrderResult = {
  insertPrescriptionOrder: ({
  id: string;
});
};

export const InsertPrescriptionOrder = {
  query: "mutation insertPrescriptionOrder($storeId: String!, $input: InsertPrescriptionOrderInput!) {\n  insertPrescriptionOrder(storeId: $storeId, input: $input) {\n    ... on PrescriptionOrderNode {\n      id\n    }\n  }\n}",
} as TypedDocument<InsertPrescriptionOrderResult, InsertPrescriptionOrderVariables>;